chore: handover to governance
Adds a test to se if we are correctly updating the ownerships and handings things over to governance.

Updated part of the deployment setup to properly hand over things. Fixed an issue with return values as we did not actually return the coinissuer but just whoever owned the fee asset - which were not the same because it was not handed over.

Co-authored-by: LHerskind <16536249+LHerskind@users.noreply.github.com>