chore: Persist validation status in archiver
Adds the _pending chain validation status_ to the archiver store, so it is persisted across node restarts. The pending chain validation status is whether the last block seen in the pending chain is valid or not, as in it having valid attestations or not.

Also, this PR fixes the condition in which it's updated: if a new block with the same number and still invalid is detected, the validation status is updated. This is tested in https://github.com/AztecProtocol/aztec-packages/pull/16836.

This PR also changes that shape of the validation result, so instead of storing a full L2 block, we only store the identifiers and attestations needed to invalidate it. It also adds serialization and static `random` methods to some adjacent classes.

Co-authored-by: PhilWindle <philip.windle@gmail.com>
Co-authored-by: Santiago Palladino <santiago@aztec-labs.com>