refactor!: nuking Aztec Node proxy methods from PXE
Having these methods exposed on PXE is a relic of a past. PXE JSON RPC server will soon be nuked and we will no longer pass a PXE_URL around and we will just pass NODE_URL.

Note that in some places this results in a worse devex as now it's necessary to pass on the input both AZTEC_NODE_URL and PXE_URL. This is only temporary so would not worry about this.

## Follow-up work
In this PR I don't touch the `PXE` proxy methods that were exposed on `Wallet` as well as it's not yet resolved what to do with those.

I also didn't yet touch `getCurrentBaseFees` as that is used by `Wallet` implementations and that would then force me to pass `AztecNode` on the input of those implementations which is a huge change. For this reason plan on doing that in a PR up the stack.