refactor!: nuking remaining Aztec Node proxy methods from PXE
Continuation of the work from the PR down the sack and nuking the remaining Aztec Node proxy methods from PXE. These are the changes:
1. Dropepd `getCurrentBaseFee`,
2. moved `getTxReceipt` on `Wallet`,
3. drop `sendTx` and moved the implementation to `BaseWallet`,
4. dropped `getNodeInfo` (this was super ugly because we had essentially a copy of implementation from AztecNode in PXE).

The cluttering of wallet inputs is only temporary because PXE is supposed to be a lib used by the wallet and hence the wallet itself should "set it up".

This makes me realize that the PXEService naming is stale as it's transforming to not be a service. Will rename that once I nuke the PXE JSON RPC server.