refactor!: nuking no-longer necessary ValueNote utils
Fixes #16879

We had utils in the `value_note` package that through time and all the API refactors just became unnecessary: they were used only in our testing contracts and I don't think it was a good place for those in the first place.

For this reason I nuke those in this PR.

This is admittedly a very low-priority task but I had a bit of spare time so decided to just tackle that.