feat: Remove estimate gas
The estimate gas call was polluting our intefaces, while the same results could be achieved by setting a combination of options in our simulation interface. This PR does just that.

It also removes the forced fee estimation process when doing `prove`. When requesting a proof, the wallet should have already decided what `GasLimits` its going to use, which is a process that does not depend on the app at all. The ability to estimate fees during simulations requested by the app is kept so the information can be shown in a GUI, for example.

Co-authored-by: Gregorio Juliana <gregojquiros@gmail.com>
Co-authored-by: thunkar <gregojquiros@gmail.com>