fix: Execute old slashing round if skipped during first round
While we keep slashing rounds for a long time (lifetime in rounds is
currently defined as 100), we only attempt to execute them during the
first round they become executable.

If for whatever reason they don't get executed, they are just forgotten
and never actually triggered.

This PR adds a new config `slashExecuteRoundsLookBack` (defaults to 4)
with how many execution rounds to look back from the latest executable
round to see if there was any round pending execution. Each round is
checked in sequence, so setting this value too high can introduce
performance issues. Setting this value to zero keeps the same behaviour
as we have today.

This PR also fixes another issue: we were re-checking if a round was
executable based on the isReadyToExecute flag returned from the
contract. However, that flag was computed based on the current at the
time of the call, and not based on the time in which the tx would land.
This meant that we always failed to execute the slash payload on the
first slot of a round.